Extract ToastNotifications - clean up top-level render
